+++ src/main/kotlin/com/hukuta94/pathnodebuilder/pipeline/filter/calculator/DistanceMatrixData.kt
@@ -5,5 +5,5 @@ import com.hukuta94.pathnodebuilder.pipeline.filter.overwatch.Vector
 data class DistanceMatrixData(
     val positions: List<Vector>,
     val connections: List<IntArray>,
-    val matrix: List<IntArray>
+    val matrix: List<DoubleArray>
 )
